Remove cert of incorporation & exemption letter
The more I think about it, the more I'm concerned that these 2 docs are the primary documents that are used to prove the organization's existence and open accounts in its name. I'm not 100% comfortable having them publicly viewable so that anyone could grab them and try to open an account or something. I'm open to leaving them if people don't agree but I wanted to raise the risk at least. We can always provide them if someone asks and has a reason for wanting to see them.