Create README.md
In talking with Asheesh, it became clear that we need to frame this section more explicitly.  It's good to have it for transparency reasons, but best to make it clear what people really need to read and what is optional if they're curious.

I'll add this page to the mkdocs menu and direct commit that update once this pull request has been approved.