Use relative URL paths for fundraising links
Links to the invoice requests, PayPal buttons, and Employer Match pages
on Fundraising Overview were broken (brought you to a GitHub 404 page).
The problem was that those URLs were missing the "docs" part after
"master", which is where those docs are located.

These changes fix that by using relative paths to get to those pages.
This means that if someone viewed the docs on GitHub, it'll link to the
right file on GitHub. If they viewed it on operations.bridgefoundry.org,
then it'll link to the right page on the website.