Update cash advance option to direct pay option
We originally were told that we couldn't do direct pay for grants to independent groups, but then it was clarified that would be okay. We had the direct pay option listed before we were really ready to implement because it seemed like that was the only way to help out groups who couldn't afford reimbursement. Since direct pay is okay after all and we want to hold off on implementing cash advances until we have a full policy in place, I'm submitting this change to list direct pay as an option instead of cash advance.